pyproject.toml
[tool.poetry]
name = "se-335-01-agent-client"
version = "0.1.0"
description = ""
authors = ["Orkun Kurul <orkunkurul@webinen.com>"]
readme = "README.md"
packages = [{include = "se_335_01_agent_client"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
platformdirs = "^4.2.1"
pyjwt = "^2.8.0"
websocket-client = "^1.8.0"
cryptography = "^42.0.7"
keyring = "^25.2.1"
browser-cookie3 = "^0.19.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
